Hooking up mass to UI

+++ Assets/Scripts/Helpers/UIHelpers.cs
@@ -3,14 +3,14 @@ using System.Collections;
 
 public class UIHelpers {
 
-    public static IEnumerator TweenTextNumberValueCoroutine( UnityEngine.UI.Text textField, int from, int to, float time, string format = "N0" )
+    public static IEnumerator TweenTextNumberValueCoroutine( UnityEngine.UI.Text textField, float from, float to, float time, string format = "N0" )
     {
-        int diff = to - from;
+        float diff = to - from;
         float startTime = Time.time;
         
         while( (Time.time - startTime) < time )
         {
-            int temp = (int)Mathf.Lerp(from, to, (Time.time - startTime) / time);
+            float temp = Mathf.Lerp(from, to, (Time.time - startTime) / time);
             if( Mathf.Abs( to - temp ) < Mathf.Abs( diff ) )
             {
                 textField.text = temp.ToString(format);
